refactor: merge `deno_url` into `deno_web`
