+++ ext/fetch/22_body.js
@@ -41,7 +41,7 @@ import * as webidl from "ext:deno_webidl/00_webidl.js";
 import {
   parseUrlEncoded,
   URLSearchParamsPrototype,
-} from "ext:deno_url/00_url.js";
+} from "ext:deno_web/00_url.js";
 import {
   formDataFromEntries,
   FormDataPrototype,
