+++ ext/web/09_file.js
@@ -58,7 +58,7 @@ const {
 
 import * as webidl from "ext:deno_webidl/00_webidl.js";
 import { ReadableStream } from "./06_streams.js";
-import { URL } from "ext:deno_url/00_url.js";
+import { URL } from "ext:deno_web/00_url.js";
 import { createFilteredInspectProxy } from "ext:deno_console/01_console.js";
 
 // TODO(lucacasonato): this needs to not be hardcoded and instead depend on
