+++ ext/web/Cargo.toml
@@ -26,12 +26,12 @@ futures.workspace = true
 serde.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
+urlpattern.workspace = true
 uuid = { workspace = true, features = ["serde"] }
 
 [dev-dependencies]
 deno_bench_util.workspace = true
 deno_console.workspace = true
-deno_url.workspace = true
 deno_webidl.workspace = true
 
 [[bench]]
@@ -40,4 +40,8 @@ harness = false
 
 [[bench]]
 name = "timers_ops"
+harness = false
+
+[[bench]]
+name = "url_ops"
 harness = false
