+++ ext/web/lib.rs
@@ -5,6 +5,8 @@ mod compression;
 mod message_port;
 mod stream_resource;
 mod timers;
+mod url;
+mod urlpattern;
 
 use std::borrow::Cow;
 use std::cell::RefCell;
@@ -53,7 +55,7 @@ use crate::timers::op_now;
 use crate::timers::op_time_origin;
 
 deno_core::extension!(deno_web,
-  deps = [ deno_webidl, deno_console, deno_url ],
+  deps = [ deno_webidl, deno_console ],
   parameters = [P: TimersPermission],
   ops = [
     op_base64_decode,
@@ -91,6 +93,14 @@ deno_core::extension!(deno_web,
     stream_resource::op_readable_stream_resource_write_sync,
     stream_resource::op_readable_stream_resource_close,
     stream_resource::op_readable_stream_resource_await_close,
+    url::op_url_reparse,
+    url::op_url_parse,
+    url::op_url_get_serialization,
+    url::op_url_parse_with_base,
+    url::op_url_parse_search_params,
+    url::op_url_stringify_search_params,
+    urlpattern::op_urlpattern_parse,
+    urlpattern::op_urlpattern_process_match_input
   ],
   esm = [
     "00_infra.js",
@@ -111,6 +121,8 @@ deno_core::extension!(deno_web,
     "14_compression.js",
     "15_performance.js",
     "16_image_data.js",
+    "00_url.js",
+    "01_urlpattern.js"
   ],
   lazy_loaded_esm = [ "webtransport.js" ],
   options = {
