+++ ext/websocket/lib.rs
@@ -870,7 +870,7 @@ pub async fn op_ws_next_event(
 }
 
 deno_core::extension!(deno_websocket,
-  deps = [ deno_url, deno_webidl ],
+  deps = [ deno_web, deno_webidl ],
   parameters = [P: WebSocketPermissions],
   ops = [
     op_ws_check_permission_and_cancel_handle<P>,
