+++ runtime/js/98_global_scope_shared.js
@@ -12,8 +12,8 @@ import * as compression from "ext:deno_web/14_compression.js";
 import * as worker from "ext:runtime/11_workers.js";
 import * as performance from "ext:deno_web/15_performance.js";
 import * as crypto from "ext:deno_crypto/00_crypto.js";
-import * as url from "ext:deno_url/00_url.js";
-import * as urlPattern from "ext:deno_url/01_urlpattern.js";
+import * as url from "ext:deno_web/00_url.js";
+import * as urlPattern from "ext:deno_web/01_urlpattern.js";
 import * as headers from "ext:deno_fetch/20_headers.js";
 import * as streams from "ext:deno_web/06_streams.js";
 import * as fileReader from "ext:deno_web/10_filereader.js";
