+++ tools/core_import_map.json
@@ -673,8 +673,8 @@
     "ext:deno_process/40_process.js": "../ext/process/40_process.js",
     "ext:deno_telemetry/telemetry.ts": "../ext/deno_telemetry/telemetry.ts",
     "ext:deno_telemetry/util.ts": "../ext/deno_telemetry/util.ts",
-    "ext:deno_url/00_url.js": "../ext/url/00_url.js",
-    "ext:deno_url/01_urlpattern.js": "../ext/url/01_urlpattern.js",
+    "ext:deno_web/00_url.js": "../ext/web/00_url.js",
+    "ext:deno_web/01_urlpattern.js": "../ext/web/01_urlpattern.js",
     "ext:deno_web/00_infra.js": "../ext/web/00_infra.js",
     "ext:deno_web/01_dom_exception.js": "../ext/web/01_dom_exception.js",
     "ext:deno_web/01_mimesniff.js": "../ext/web/01_mimesniff.js",
